Move upload file logic to api_v6 folder
