+++ server/service/src/sync/file_synchroniser.rs
@@ -154,10 +154,15 @@ impl FileSynchroniser {
                     ..file.clone()
                 })?;
 
-                let result = self.try_uploading_file(file).await;
+                let result = self
+                    .sync_api_v6
+                    .upload_file(&self.static_file_service, file)
+                    .await;
 
-                let bytes_uploaded = match result {
-                    Ok(bytes_uploaded) => bytes_uploaded,
+                match result {
+                    Ok(_) => {
+                        log::debug!("File uploaded successfully");
+                    }
                     Err(err) => {
                         log::error!("Error uploading file: {:#?}", err);
 
@@ -203,9 +208,9 @@ impl FileSynchroniser {
                     }
                 };
 
-                // Update database to record the chunk has been uploaded
+                // Update database to record the file has been uploaded
                 sync_file_repo.update_status(&SyncFileReferenceRow {
-                    uploaded_bytes: file.uploaded_bytes + bytes_uploaded,
+                    uploaded_bytes: file.total_bytes, // We always upload the whole file in one go
                     status: SyncFileStatus::Done,
                     error: None,
                     ..file.clone()
@@ -219,118 +224,5 @@ impl FileSynchroniser {
         let num_of_files = files.len();
 
         Ok(num_of_files)
-    }
-
-    // Move to sync_api_v6
-    async fn try_uploading_file(
-        &self,
-        sync_file_reference_row: &SyncFileReferenceRow,
-    ) -> Result<i32, FileSyncError> {
-        // Get file path
-        let file = self
-            .static_file_service
-            .find_file(
-                &sync_file_reference_row.id,
-                StaticFileCategory::SyncFile(
-                    sync_file_reference_row.table_name.to_owned(),
-                    sync_file_reference_row.record_id.to_owned(),
-                ),
-            )
-            .map_err(|err| {
-                log::error!("Error from static_file_service: {:#?}", err);
-                FileSyncError::CantFindFile("Error from static_file_service".to_string())
-            })?;
-        let file = match file {
-            Some(file) => file,
-            None => {
-                return Err(FileSyncError::CantFindFile(
-                    "File doesn't exist in static_file_service".to_string(),
-                ))
-            }
-        };
-
-        let mut file_handle = std::fs::File::open(file.path.clone()).map_err(|err| {
-            log::error!("Error opening file: {:#?}", err);
-            FileSyncError::StdIoError(err)
-        })?;
-
-        // Read the file into memory (ideally could be a stream or something, and upload the file in chunk so we can stop quickly when sync starts/stops/pauses)
-        let mut file_bytes = Vec::new();
-        file_handle.read_to_end(&mut file_bytes).map_err(|err| {
-            log::error!("Error reading file: {:#?}", err);
-            FileSyncError::StdIoError(err)
-        })?;
-
-        let file_upload_part = reqwest::multipart::Part::bytes(file_bytes)
-            .file_name(sync_file_reference_row.file_name.clone());
-
-        let sync_api_settings_json =
-            serde_json::to_string(&self.sync_api_settings).map_err(|err| {
-                log::error!("Error serializing sync_api_settings: {:#?}", err);
-                FileSyncError::UploadError(UploadError::Other(
-                    "Error serializing sync_api_settings".to_string(),
-                ))
-            })?;
-        let sync_settings_part = reqwest::multipart::Part::text(sync_api_settings_json)
-            .mime_str("application/json")
-            .unwrap(); // TODO fix unwrap
-
-        let form = multipart::Form::new()
-            .part("sync_v5_settings", sync_settings_part)
-            .part("file", file_upload_part);
-
-        // Calculate url for upload
-        let url = self
-            .settings
-            .file_upload_url()
-            .join(&sync_file_reference_row.id)
-            .map_err(|err| {
-                log::error!("Error creating url: {:#?}", err);
-                FileSyncError::UploadError(UploadError::Other(
-                    "Error creating url for id".to_string(),
-                ))
-            })?;
-        log::info!("Uploading {} to {}", sync_file_reference_row.file_name, url);
-
-        // Upload file
-        let request = self.client.put(url).multipart(form);
-        println!("request: {:#?}", request);
-        let request = request.send().await;
-        match request {
-            Ok(response) => {
-                if response.status().is_success() {
-                    log::info!("File {} uploaded successfully", sync_file_reference_row.id);
-                } else {
-                    let status = response.status();
-                    let text = response.text().await.unwrap_or_default();
-
-                    log::error!(
-                        "Error uploading file {} - {} : {:#?}",
-                        sync_file_reference_row.id,
-                        status,
-                        text
-                    );
-
-                    if status == reqwest::StatusCode::NOT_FOUND {
-                        return Err(FileSyncError::UploadError(UploadError::NotFound));
-                    }
-
-                    return Err(FileSyncError::UploadError(UploadError::Other(format!(
-                        "{}:{}",
-                        status, text
-                    ))));
-                }
-            }
-            Err(err) => {
-                log::error!("Error uploading file: {:#?}", err);
-                if err.is_connect() {
-                    return Err(FileSyncError::UploadError(UploadError::ConnectionError));
-                }
-                return Err(FileSyncError::ReqwestError(err));
-            }
-        }
-
-        let bytes_uploaded = sync_file_reference_row.total_bytes as i32; // Currently just uploading the whole file
-        Ok(bytes_uploaded)
     }
 }
